Initial commit; MVP
Allows for simple command-line batch file rename

Takes two arguments. First is the string to search for, and the second
is the string to replace with. Searches within current directory (the
directory in which the program is executed).

Next step would be to add an option to search folders recursively, and
to add a GUI